code/labs/requirements.txt
duckdb
deepdiff
numpy
pandas
rdflib
networkx
matplotlib
jellyfish
python-dotenv